requirements.txt
Flask==1.1.2
Flask-MySQL==1.5.2
mysql-client==0.0.1

